c++17, pub/pvt header, test cmake c++ std level
- enable c++17 std::filesystem support and test cases
- split header into public/private parts
- cmake for test cases now supports optional
  c++ standard level param
