+++ src/openvino/OpenVINO.cpp
@@ -10,6 +10,7 @@
 
 #include "BlobReader.hpp"
 #include "spdlog/spdlog.h"
+#include "utility/spdlog-fmt.hpp"
 
 namespace dai {
